+++ lib-network/include/linux/network.h
@@ -212,7 +212,7 @@ public:
 
 	int32_t TcpBegin(uint16_t nLocalPort);
 	uint16_t TcpRead(const int32_t nHandle, const uint8_t **ppBuffer, uint32_t &HandleConnection);
-	void TcpWrite(const int32_t nHandle, const uint8_t *pBuffer, uint16_t nLength, const uint32_t HandleConnection);
+	void TcpWrite(const int32_t nHandle, const uint8_t *pBuffer, uint32_t nLength, const uint32_t HandleConnection);
 	int32_t TcpEnd(const int32_t nHandle);
 
 private:
